move vespa-logfmt command
