+++ client/go/script-utils/logfmt/cmd.go
