+++ client/go/script-utils/logfmt/handleline.go
