+++ client/go/script-utils/logfmt/internal.go
