+++ client/go/script-utils/logfmt/internal_test.go
