+++ client/go/script-utils/logfmt/regexflag.go
