+++ client/go/script-utils/logfmt/regexflag_test.go
