+++ client/go/script-utils/logfmt/runlogfmt.go
