+++ client/go/script-utils/logfmt/tail.go
